v16-e2e/mcp-servers/complaint-management-mcp/requirements.txt
fastmcp==2.10.6
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
faker>=20.0.0
colorama>=0.4.6
art>=6.1
